seks

+++ src/features/events/eventsService.js
@@ -18,15 +18,11 @@ export const eventsService = {
     return data
   },
   async registerUser(eventId, userId) {
-    const { data } = await apiClient.post(
-      `/api/events/${eventId}/registrations/users/${userId}`
-    )
+    const { data } = await apiClient.post(`/api/events/${eventId}/register/${userId}`)
     return data
   },
   async removeUser(eventId, userId) {
-    const { data } = await apiClient.delete(
-      `/api/events/${eventId}/registrations/users/${userId}`
-    )
+    const { data } = await apiClient.delete(`/api/events/${eventId}/register/${userId}`)
     return data
   },
 }
